requirements.txt
requests>=2.31.0
sodapy>=1.6.1
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gspread>=5.12.0
oauth2client>=4.1.3
pandas>=2.1.0
python-dotenv>=1.0.0
